requirements.txt
langgraph>=0.2.0
langchain-google-genai>=1.0.0
google-cloud-bigquery>=3.13.0
pandas>=2.0.0
python-dotenv>=1.0.0 
langchain_core>=0.3.0
db-dtypes==1.2.0
pydantic>=2.10.1
langchain_community >= 0.3.27
langchain == 1.0.4
matplotlib == 3.10.7
streamlit == 1.51.0
Pillow == 12.0.0